+++ src/react-notifications.ts
@@ -1,3 +1,3 @@
-export * from './lib/Container'
+export * from './lib/NotificationContainer'
 export * from './lib/Notification'
 export * from './lib/state'